Rewrite all the makefiles - now EASTL is built into a library
I plan on adding more examples, so building into a library and being able
to compile multiple examples makes sense, and it's nice to be able to take
the compiled library and link it conveniently into your own projects as
you wish.

Also rename the only example since there will be more soon.